+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `taxValue` on `product-selling-price` and `product-list-price`
+
 ## [1.12.0] - 2021-01-06
 ### Added
 - Support for custom classes.